move stack add_node out
